refactor(frontend): extract match editor internals
